Add reminder to create index for custom indexer
And various small fixes.
This page refers to AT, should be converted to DX for Plone 5 in another commit.
Also, the warning about catalog.xml GS import step clearing indexes is no longer relevant for Plone 5.  Technically, it could be removed, but it might still be helpful for someone switching between 4 and 5 projects.